+++ src/Sulu/Bundle/ContentBundle/Resources/config/serializer/Document.BasePageDocument.xml
@@ -22,7 +22,7 @@
         <property name="path" type="string" />
         <property name="extensions" type="Sulu\Component\Content\Document\Extension\ExtensionContainer" />
 
-        <property name="structure" type="Sulu\Component\Content\Document\Property\PropertyContainer" />
+        <property name="structure" type="Sulu\Component\Content\Document\Structure\Structure" />
         <property name="children" exclude="true" type="Sulu\Component\DocumentManager\Collection\ChildrenCollection" />
         <property name="parent" exclude="true" />
     </class>
